Finalize blog post rating

+++ service/src/main/java/com/gdevxy/blog/service/contentful/blogpost/converter/BlogPostConverter.java
@@ -24,7 +24,7 @@ public class BlogPostConverter {
 
 	private final ImageConverter imageConverter;
 
-	public  BlogPost convert(PageBlogPost p, Integer rating) {
+	public  BlogPost convert(PageBlogPost p, Long rating) {
 
 		return BlogPost.builder()
 			.id(p.getSys().getId())
